+++ abc-bank/Accounts/SavingsAccount.cs
@@ -3,15 +3,15 @@
 
     public override string ReportLabel { get; } = "Savings Account";
 
-    public override double InterestEarned {
+    public override decimal InterestEarned {
       get {
         return CurrentBalance <= 1000
-          ? CalculateInterest(0.001)
-          : CalculateInterest(0.002);
+          ? CalculateInterest(0.001M)
+          : CalculateInterest(0.002M);
       }
     }
 
-    public SavingsAccount(int accountId, int customerId, double initialDeposit = 0.00)
+    public SavingsAccount(int accountId, int customerId, decimal initialDeposit = 0)
       : base(AccountType.SAVINGS, accountId, customerId, initialDeposit) { }
 
   }
